backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
pyyaml==6.0.1
pandas==2.1.3
opencv-python==4.8.1.78
mediapipe==0.10.8
ultralytics==8.0.200
numpy==1.24.3
Pillow==10.1.0
torch>=2.0.0
torchvision>=0.15.0